ApplicationController: Provide ways to register dependent Widgets
It is desirable that some widgets should be
dependent on the `ApplicationController`'s magic
information (specifically, the logged-in user).

We want to be able to add these in an asynchronous
manager so we provide the registration method
(which immediately is called back for
initialization purposes).

We also add that selfsame callback to a callback
queue that will be called when the critical datum
(i.e. the `loggedUser`) is updated.