Pairings#get_hangout_url: log params for thorny method
This method is in the middle of several callbacks
and may fail if the callback is triggered from the
wrong domain (i.e. you're using your own
development account but launched your hangout
session with different credentials).

This datum is really helpful in not wasting a ton
of time debugging.

Conflicts:
	app/controllers/pairings_controller.rb